Let update facet normal and offsets take containers as arguments

+++ TestConvexHull.cpp
@@ -295,7 +295,7 @@ int main( int argc, const char* argv[] )
   }
   else {
     if ( atoi( argv[ 1 ] ) == 0 ) {
-      size_t numOfPoints = 1e6;
+      size_t numOfPoints = size_t( 1e6 );
       size_t dimension = 3;
       if ( argc > 2 ) {
         int np = atoi( argv[ 2 ] );
@@ -314,8 +314,7 @@ int main( int argc, const char* argv[] )
     }
     else {
       testSpeedUniform_();
-      testSpeedRandom_( 1e6, 3 );
+      testSpeedRandom_( size_t( 1e6 ), 3 );
     }
   }
-
 }
